SET/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
# Note: pydantic is installed automatically by fastapi with a compatible pydantic-core version
# Do not specify pydantic version explicitly to avoid Rust compilation issues
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
pypandoc-binary==1.13
reportlab==4.2.5
pypdf==4.3.1
aiofiles==24.1.0
python-multipart==0.0.12
beautifulsoup4==4.12.3
lxml==5.3.0

# LinkedIn Scraper dependencies
playwright==1.49.0

# AI Services
openai==1.59.0

# Web Search - for DuckDuckGo fallback (when Google API is not configured)
duckduckgo-search==6.3.5  # Free search alternative to Google API